slight cleanup

+++ Content/MissileAddons/BeamCombos/Wavebuster.cs
@@ -1,6 +1,5 @@
 using System;
 using System.IO;
-using MetroidMod.Common.GlobalItems;
 using MetroidMod.Content.BeamAddons;
 using MetroidMod.Content.Projectiles;
 using MetroidMod.ID;
@@ -230,8 +229,8 @@ namespace MetroidMod.Content.MissileAddons.BeamCombos
 					{
 						//targetPos = new Vector2(mousePos.X + Main.rand.Next(-30, 31), mousePos.Y + Main.rand.Next(-30, 31));
 						targetPos = oPos + diff * range;
-						targetPos.X += (float)Main.rand.Next(-30, 31) * (Vector2.Distance(oPos, P.Center) / Max_Range);
-						targetPos.Y += (float)Main.rand.Next(-30, 31) * (Vector2.Distance(oPos, P.Center) / Max_Range);
+						targetPos.X += Main.rand.Next(-30, 31) * (Vector2.Distance(oPos, P.Center) / Max_Range);
+						targetPos.Y += Main.rand.Next(-30, 31) * (Vector2.Distance(oPos, P.Center) / Max_Range);
 					}
 				}
 
@@ -398,9 +397,9 @@ namespace MetroidMod.Content.MissileAddons.BeamCombos
 					{
 						rot = (float)Math.Atan2((pos[i].Y - pos[i - 1].Y), (pos[i].X - pos[i - 1].X)) + (float)Math.PI / 2;
 					}
-					sb.Draw(tex, pos[i] - Main.screenPosition, new Rectangle?(new Rectangle(0, y4, tex.Width, num108)), P.GetAlpha(Color.White), rot, new Vector2((float)tex.Width / 2f, (float)num108 / 2), new Vector2(scale, 1f), SpriteEffects.None, 0f);
+					sb.Draw(tex, pos[i] - Main.screenPosition, new Rectangle?(new Rectangle(0, y4, tex.Width, num108)), P.GetAlpha(Color.White), rot, new Vector2(tex.Width / 2f, (float)num108 / 2), new Vector2(scale, 1f), SpriteEffects.None, 0f);
 
-					sb.Draw(tex2, pos[i] - Main.screenPosition, new Rectangle?(new Rectangle(0, numH * Main.rand.Next(4), tex2.Width, numH)), P.GetAlpha(Color.White), rot, new Vector2((float)tex2.Width / 2, (float)numH / 2), (float)(Main.rand.Next(21) / 10), SpriteEffects.None, 0f);
+					sb.Draw(tex2, pos[i] - Main.screenPosition, new Rectangle?(new Rectangle(0, numH * Main.rand.Next(4), tex2.Width, numH)), P.GetAlpha(Color.White), rot, new Vector2((float)tex2.Width / 2, (float)numH / 2), Main.rand.Next(21) / 10, SpriteEffects.None, 0f);
 
 					Lighting.AddLight(pos[i], (MetroidMod.waveColor2.R / 255f) * P.scale, (MetroidMod.waveColor2.G / 255f) * P.scale, (MetroidMod.waveColor2.B / 255f) * P.scale);
 
@@ -422,7 +421,7 @@ namespace MetroidMod.Content.MissileAddons.BeamCombos
 				immuneTime += O.HeldItem.useTime;
 				mProjectile.Projectile.localNPCHitCooldown = O.HeldItem.useTime;
 			}
-			base.OnHitNPC(mProjectile, target, hit,damageDone);
+			base.OnHitNPC(mProjectile, target, hit, damageDone);
 		}
 		public override void SendExtraAI(BinaryWriter writer)
 		{
